images now representable as 32x32x3 vectors
two methods put in placeholder because I was just experimenting.
"shrinkImage()" rips the code from ImageTest into 1 effective function,
and "rgbVector()" takes an image name, uses shrinkImage() to shrink it,
and then returns an int[] vector of rgb values which is the 32*32 red
channel followed by the green and blue channel (in that order). There is
1 last method called test() that verifies both of these work.