added histogram() method
uses BufferedImage class to turn an image into an object, iterates
through every pixel in the image, and increments a bin in a
3-dimensional array with the corresponding (R,G,B) coordinates. Remember
that this isn't a 256x256x256 histogram; it's 8x8x8. I divide each R, G,
and B value by 32 (as 256/32=8) to decrease the range of values to fit
into the 8 bins per dimension. I DID NOT NORMALIZE THE HISTOGRAM'S BIN
VALUES.