patched up indexG to actually reset and to write results to a file
GaussianTuning.txt contains the large set of results for many C and
gamma values (110 iterations intended, maybe 85 done). GaussianTuning2
is the tuning results on Tiny Image data from a narrower range of values
with more relevant gamma values. GaussianTuning3 is going to be the
results of tuning for the histograms